Added currency management.

+++ models/src/main/java/si/fri/rso/uniborrow/cash/models/entities/CashEntity.java
@@ -17,7 +17,7 @@ public class CashEntity {
     private Integer userId;
 
     @Column(name = "current_cash")
-    private Integer currentCash;
+    private Float currentCash;
 
 }
